% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayArabic.R
\name{displayArabic}
\alias{displayArabic}
\title{Display Arabic Correctly}
\usage{
displayArabic(Arabic_string)
}
\arguments{
\item{Arabic_string}{String you want to display in ggplot}
}
\value{
A string of readable Arabic
}
\description{
This function displays Arabic characters in the correct order
using the correct connections. It calls a python file in the
DisplayArabic package that reverses the order of the Arabic
characters using the python-bidi module and called the correct
character form based on the character's position in the word
use the python arabic.reshaper module. The object the function
returns will display Arabic strings correctly in ggplot.
}
