% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_required_python.R
\name{install_bidi}
\alias{install_bidi}
\title{Install python-bidi package}
\usage{
install_bidi(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method; choice of "auto", "virtualenv", or "conda"}

\item{conda}{Path to conda executable; "auto" uses reticulate to automatically find path}
}
\value{
bidi is successfully installed
}
\description{
This function installs required bidirectional python module if the
user does not have the package installed already. This module reverses
the order of the characters.
}
