% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_required_python.R
\name{set_python_path}
\alias{set_python_path}
\title{Set python path}
\usage{
set_python_path(pth = "general")
}
\arguments{
\item{pth}{path to User's preferred python version; must be a string}
}
\value{
reticulate using the preferred python version, defaults to return of system("which python")
}
\description{
This function allows the User to set to their preferred version of
python. In particular, if the User has already installed python-bidi
or arabic-reshaper, the User can choose set the path here to where
those packages are installed.
}
